% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backseries.R
\name{add_backseries}
\alias{add_backseries}
\title{Add yearly backseries to data}
\usage{
add_backseries(data, drive_name, file_name, year, date_col = FALSE)
}
\arguments{
\item{data}{Data frame of current year data.}

\item{drive_name}{String of name of drive where backseries data is stored.}

\item{file_name}{String of csv file name.}

\item{year}{Numeric value of current year of data (YYYY).}

\item{date_col}{Boolean (TRUE or FALSE) for whether the backseries needs to
be joined onto the right of the current data using the Date column (i.e. rows
contain days of the year, not \code{Years}). Default is FALSE.}
}
\value{
Data frame.
}
\description{
Binds the backseries data below the current year data (has a
column for "Year" which is descending). Or binds the backseries to the right
of the current year data where columns of year descend to the right and the
rows are "Date" if specified.
}
\details{
If the current \code{year} of data already exists in the backseries, will
error.
}
