% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_tables.R
\name{create_unreliability_columns}
\alias{create_unreliability_columns}
\title{Creates unreliability columns to data frame}
\usage{
create_unreliability_columns(data, columns, lower = 3, upper = 19)
}
\arguments{
\item{data}{Output data frame.}

\item{columns}{Vector of quoted column names to be checked against threshold.}

\item{lower}{Integer value of lower threshold to use (inclusive).
Default is 3.}

\item{upper}{Integer value of upper threshold to use (inclusive).
Default is 19.}
}
\value{
Data frame with additional unreliability columns if applicable.
}
\description{
Checks whether any values from the specified columns contain
counts that are considered unreliable. If so, it adds unreliability columns
and populates, otherwise returns original data frame.
}
