% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_tables.R
\name{format_multi_mat_count_rate_agegrp}
\alias{format_multi_mat_count_rate_agegrp}
\title{Formats the output of \code{multi_mat_count_rate_agegrp_ew} to publishing-ready
single row of data}
\usage{
format_multi_mat_count_rate_agegrp(data, year)
}
\arguments{
\item{data}{Dataframe output of \code{multi_mat_count_rate_agegrp_ew}.}

\item{year}{Numeric or string value of year of data (YYYY).}
}
\value{
Single dataframe row of data for publication.
}
\description{
Formats the output of \code{multi_mat_count_rate_agegrp_ew} to publishing-ready
single row of data
}
