% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_imports.R
\name{import_births_ew}
\alias{import_births_ew}
\title{Import births data set for chosen year}
\usage{
import_births_ew(sql_server, database, year)
}
\arguments{
\item{sql_server}{Name of SQL server, as a character string.}

\item{database}{Name of database, as a character string.}

\item{year}{4-digit year of data, as integer.}
}
\value{
Data frame of births data.
}
\description{
Import births data set for chosen year
}
