% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_imports.R
\name{import_nspl}
\alias{import_nspl}
\title{Import National Statistics Postcode Lookup (NSPL)}
\usage{
import_nspl(sql_server, database, met_county_colname)
}
\arguments{
\item{sql_server}{Name of SQL server, as a character string.}

\item{database}{Name of database, as a character string.}

\item{met_county_colname}{Column name for metropolitan county codes in NSPL
data set, which varies in each version, as a character string.}
}
\value{
Data frame of NSPL data.
}
\description{
Import National Statistics Postcode Lookup (NSPL)
}
