% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{join_births_nspl}
\alias{join_births_nspl}
\title{Join births and NSPL data}
\usage{
join_births_nspl(births_df, nspl)
}
\arguments{
\item{births_df}{Births data frame containing \code{pcdrm} column.}

\item{nspl}{NSPL data frame.}
}
\value{
Births data frame with NSPL information.
}
\description{
Removes the space from the births postcode column (pcdrm) and
then left joins the NSPL onto the births dataframe.
}
