% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tables.R
\name{multi_mat_count_rate_agegrp_ew}
\alias{multi_mat_count_rate_agegrp_ew}
\title{Multiple birth counts and rate by mothers age group (England and Wales)}
\usage{
multi_mat_count_rate_agegrp_ew(data)
}
\arguments{
\item{data}{Processed birth dataframe (registrations) for a single year.}
}
\value{
Dataframe of multiple births and rates per mothers age group for a
single year.
}
\description{
Uses processed births registrations for England and Wales to
create counts of multiple births and multibirth rate by mothers age group.
}
