% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{process_ew_notif}
\alias{process_ew_notif}
\title{Process births data from England and Wales}
\usage{
process_ew_notif(df, nspl, gest_col)
}
\arguments{
\item{df}{Births data frame from England and Wales, registrations or
notifications.}

\item{nspl}{NSPL data frame.}

\item{gest_col}{Column name with gestation time, either "gestatn" or
"gest10", as a string.}
}
\value{
Births data frame with added geography columns, age (in years) for
mothers, age groupings (agegrp) for mothers, gestational time groupings
(gest) and number of previous live-born children.
}
\description{
Apply processing to both births registrations and notifications
data from England and Wales. Joins NSPL codes, changes mother's age to 2
digits and add gestation weeks groupings and number of previous live-born
children.
}
