% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{process_nspl}
\alias{process_nspl}
\title{Process NSPL data}
\usage{
process_nspl(nspl, met_county_colname)
}
\arguments{
\item{nspl}{NSPL data frame.}

\item{met_county_colname}{Column name for Metropolitan County codes in NSPL,
which varies in each NSPL data version, as a character string.}
}
\value{
Processed NSPL data frame.
}
\description{
Process National Statistics Postcode Location (NSPL) data.
Rename column containing Metropolitan County codes to "mcounty", remove space
from postcodes column and create a joint local authority unitary authority
(laua) code for Cornwall and Isle of Scilly.
}
