% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_tables.R
\name{unreliable_rate_mat_mulbth}
\alias{unreliable_rate_mat_mulbth}
\title{Add unreliable columns to maternities with multiple births rate}
\usage{
unreliable_rate_mat_mulbth(data)
}
\arguments{
\item{data}{Data frame.}
}
\value{
Data frame with unreliability indicator columns if applicable.
}
\description{
Add unreliable columns to maternities with multiple births rate
}
